+++ .gitignore
@@ -1,0 +1,10 @@
+# env
+.env
+Env
+env
+
+# cache
+*.sqlite3
+__pycache__
+**/__pycache__
+**/**/__pycache__
