+++ README.md
@@ -1,0 +1,35 @@
+# Estimasi Harga Mobil Bekas
+
+## Instalasi
+
+#### 1. Menyiapkan Virtual Environment
+
+```sh
+py -m venv env 
+```
+
+#### 2. Masuk Kedalam Virtual Environment
+
+Windows:
+
+```sh
+env/Scripts/activate
+```
+
+Unix:
+
+```sh
+source env/Scripts/activate
+```
+
+#### 3. Mengunduh Semua Library
+
+```sh
+py setup.py
+```
+
+#### 4. Menjalankan Server
+
+```sh
+py manage.py runserver
+```
