+++ django_app/models/mobil_model.py
