+++ django_app/models/model_knn.pkl
