+++ django_app/urls.py
@@ -13,11 +13,11 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
+
 from django.urls import path
 from . import views
+
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', views.estimasi_harga),
-    
+  path('', views.home),
+  path('estimate/', views.estimasi_harga)  
 ]
