+++ django_app/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for belajar project.
+WSGI config for django_app project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'belajar.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_app.settings')
 
 application = get_wsgi_application()
