+++ installed_packages.txt
@@ -1,0 +1,9 @@
+asgiref==3.7.2
+Django==4.2.5
+joblib==1.3.2
+numpy==1.26.0
+scikit-learn==1.3.1
+scipy==1.11.3
+sqlparse==0.4.4
+threadpoolctl==3.2.0
+tzdata==2023.3
