+++ manage.py
@@ -3,19 +3,18 @@
 import os
 import sys
 
-
 def main():
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'belajar.settings')
-    try:
-        from django.core.management import execute_from_command_line
-    except ImportError as exc:
-        raise ImportError(
-            "Couldn't import Django. Are you sure it's installed and "
-            "available on your PYTHONPATH environment variable? Did you "
-            "forget to activate a virtual environment?"
-        ) from exc
-    execute_from_command_line(sys.argv)
+  os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_app.settings')
 
+  try:
+    from django.core.management import execute_from_command_line
+  except ImportError as exc:
+    raise ImportError(
+      "Couldn't import Django. Are you sure it's installed and "
+      "available on your PYTHONPATH environment variable? Did you "
+      "forget to activate a virtual environment?"
+    ) from exc
+  execute_from_command_line(sys.argv)
 
 if __name__ == '__main__':
-    main()
+  main()
