+++ templates/estimasi_harga.html
@@ -1,21 +1,29 @@
+{% load static %}
+
 <!DOCTYPE html>
-<html>
-<head>
+<html lang="id">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="IE=7" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+
+    <link rel="icon" href="{% static 'images/icon.png' %}" />
+
     <title>Estimasi Harga Mobil Bekas</title>
-</head>
-<body>
+  </head>
+  <body>
     <h1>Estimasi Harga Mobil Bekas</h1>
     <form method="post">
-        {% csrf_token %}
-        {{ form.as_p }}
-        <input type="submit" value="Estimasi Harga">
+      {% csrf_token %} {{ form.as_p }}
+      <input type="submit" value="Estimasi Harga" />
     </form>
-    
+
     {% if estimasi %}
     <h2>Estimasi Harga:</h2>
-    <p> {{ estimasi }}</p>
-    
+    <p>{{ estimasi }}</p>
+
     <p>atau jika dirupiahkan yaitu: Rp {{ estimasi_rupiah }} Juta</p>
     {% endif %}
-</body>
+  </body>
 </html>
