+++ templates/home.html
@@ -1,0 +1,1 @@
+<h1>Home</h1>
